Sign Language Detection
As part of the Erciyes University Design Project, an image processing study was carried out to detect letters expressed in sign language from images within a real-time video stream. In the project, a dataset created using American Sign Language (ASL) as a reference was processed with dynamic scaling and classification models. The detection of letters expressed in sign language was successfully achieved using a model developed with the Keras library.